+++ cynic-codegen/tests/use-schema.rs
@@ -10,7 +10,8 @@ use cynic_codegen::use_schema::{use_schema, QueryDslParams};
 #[rstest(schema_file => [
     "graphql.jobs.graphql",
     "books.graphql",
-    "starwars.schema.graphql"
+    "starwars.schema.graphql",
+    "test_cases.graphql"
 ])]
 fn snapshot_use_schema(schema_file: &str) {
     let schema_path = PathBuf::from("../schemas/").join(schema_file);